delete batches, added location spinner, etc

+++ app/src/main/java/com/example/slfastener/model/grnlineitemmain/ResponseObject.kt
@@ -24,6 +24,6 @@ data class ResponseObject(
     val poNumber: String,
     val poQty: Int,
     val pouom: String,
-    val retBalQty: Int,
+    val retBalQty: Double,
     val unitPrice: Any
 )
